flask app/requirements.txt
Flask
CodeforcesApiPy

